refactor: 수업 교환 훅 분리
